code-samples/python/requirements.txt
# AI and Machine Learning
openai>=1.0.0
transformers>=4.21.0
langchain>=0.1.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
pyyaml>=6.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Web frameworks (optional)
flask>=2.2.0
streamlit>=1.25.0

# File processing
pypdf2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.10